refactor(tokenizer): split message_preprocess tests into child module (≤500 LoC)
The developer-role test additions pushed message_preprocess.rs to 533
lines, over the 500-LoC-per-file CI cap. Move the `#[cfg(test)] mod tests`
body into `message_preprocess/tests.rs` (child module — identical
`super`/`super::super` paths, no test logic change). Parent drops to 216
lines, tests file 320; both under the cap. All 16 lib tests still pass.
